Make changes to work with tree
Did this to cover all children and not just first level

This also has the benefit now that the monthly totals
come as part of the calculation in the base account.

For ex: base account for expenses has all totals of
its children. Which would be the same as old Totals row

This however doesn't have the filtering. For now, clients will
do that on their own.