Print transactions
Print transactions as json as well. Chose the monthly model for
this so that there is less reading and writing going on. The data
won't change much so it's beneficial to do less. And to do that,
there's a config which allows writing transactions for only
current and previous month. Might expand that to all reports if
this works well.