Fix account full names empty some times
There was an issue with running this on a Raspberry Pi3,
where for some accounts only, the parent account is null
when reading transactions. The same code runs okay on the Mac
but the same accounts are null every time on Raspberry Pi3.

After a lot of trial, ended up creating a dict up front with
account id and their full names and use it if full name was
empty.